[add]: finished troubleshooting cluster
