+++ apps/api/Dockerfile
@@ -1,0 +1,6 @@
+FROM scratch
+MAINTAINER DavyJ0nes <davy.jones@me.com>
+
+ADD davyj-ao-training-api /
+EXPOSE 8000
+CMD ["./davyj-ao-training-api"]
