+++ apps/api/env-example
@@ -1,0 +1,1 @@
+export ECR_REPO=<<URL OF ECR REPO>>
