--- apps/fred/Dockerfile
@@ -1,6 +1,0 @@
-FROM scratch
-MAINTAINER DavyJ0nes <davy.jones@me.com>
-
-ADD fred /
-EXPOSE 8080
-CMD ["./fred"]
