+++ apps/front_end/Dockerfile
@@ -1,0 +1,6 @@
+FROM scratch
+MAINTAINER DavyJ0nes <davy.jones@me.com>
+
+ADD davyj-ao-training-frontend /
+EXPOSE 8080
+CMD ["./davyj-ao-training-frontend"]
