--- apps/json_jill/Dockerfile
@@ -1,6 +1,0 @@
-FROM scratch
-MAINTAINER DavyJ0nes <davy.jones@me.com>
-
-ADD json_jill /
-EXPOSE 8000
-CMD ["./json_jill"]
