--- apps/steve/Dockerfile
@@ -1,6 +1,0 @@
-FROM scratch
-MAINTAINER DavyJ0nes <davy.jones@me.com>
-
-ADD steve /
-EXPOSE 8080
-CMD ["./steve"]
