+++ terraform/modules/ecr/variables.tf
@@ -1,10 +1,10 @@
 # ECR Module Variables
 # DavyJ0nes 2017
 
-variable "env" {
-  description = "The Environment name of the Repository"
-}
-
 variable "prefix" {
   description = "Naming prefix to use"
+}
+
+variable "app" {
+  description = "The application name"
 }
