+++ terraform/modules/ecs/service/ecs_service.tf
@@ -1,0 +1,4 @@
+# ECS Service Module
+# DavyJ0nes 2017
+#----------- SET UP SERVICE ----------#
+
